requirements.txt
django==4.2
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==7.2.0
phonenumbers==8.13.39
geographiclib==2.0
geopy==2.4.1
requests==2.32.3
djangorestframework==3.15.2
python-dotenv==1.0.1